+++ packages/angular-cli/tasks/serve-webpack.ts
@@ -19,8 +19,12 @@ export default Task.extend({
     let webpackCompiler: any;
 
     let config = new NgCliWebpackConfig(
-      this.project, commandOptions.target,
-      commandOptions.environment
+      this.project,
+      commandOptions.target,
+      commandOptions.environment,
+      undefined,
+      undefined,
+      commandOptions.aot
     ).config;
 
     // This allows for live reload of page when changes are made to repo.
@@ -54,7 +58,10 @@ export default Task.extend({
       historyApiFallback: true,
       stats: webpackDevServerOutputOptions,
       inline: true,
-      proxy: proxyConfig
+      proxy: proxyConfig,
+      watchOptions: {
+        poll: CliConfig.fromProject().config.defaults.poll
+      }
     };
 
     ui.writeLine(chalk.green(oneLine`