+++ packages/webpack/src/index.ts
@@ -1,0 +1,4 @@
+import 'reflect-metadata';
+
+export * from './plugin'
+export {ngcLoader as default} from './loader'